+++ src/Silverback.Integration.MQTT/Messaging/Broker/Mqtt/MqttNetClientFactory.cs
@@ -3,6 +3,7 @@
 
 using MQTTnet;
 using MQTTnet.Client;
+using MQTTnet.Diagnostics;
 
 namespace Silverback.Messaging.Broker.Mqtt
 {
@@ -11,9 +12,22 @@ namespace Silverback.Messaging.Broker.Mqtt
     /// </summary>
     public class MqttNetClientFactory : IMqttNetClientFactory
     {
+        private readonly IMqttNetLogger _mqttNetLogger;
+
         private readonly MqttFactory _factory = new();
 
+        /// <summary>
+        ///     Initializes a new instance of the <see cref="MqttNetClientFactory" /> class.
+        /// </summary>
+        /// <param name="mqttNetLogger">
+        ///     The <see cref="IMqttNetLogger" />.
+        /// </param>
+        public MqttNetClientFactory(IMqttNetLogger mqttNetLogger)
+        {
+            _mqttNetLogger = mqttNetLogger;
+        }
+
         /// <inheritdoc cref="IMqttNetClientFactory.CreateClient" />
-        public IMqttClient CreateClient() => _factory.CreateMqttClient();
+        public IMqttClient CreateClient() => _factory.CreateMqttClient(_mqttNetLogger);
     }
 }